requirements.txt
# streamlit
# openai>=1.31.0
# python-dotenv
# python-docx
# sentence-transformers>=2.2.2
# torch>=2.1.0
# PyPDF2

# streamlit
# openai
# python-dotenv
# python-docx
# sentence-transformers
# torch
# PyPDF2

streamlit
openai==1.3.5
python-dotenv
python-docx
sentence-transformers==2.3.1
torch==2.2.0 --extra-index-url https://download.pytorch.org/whl/cpu
PyPDF2
